// src/game/data/enums.ts
export const enum ESkillFeatureType {
  NONE = 0,
  ENABLE_NOTIFY = 1,
  ENABLE_DIR_CHANGE = 2,
  CHANGE_MOVE_DIST = 3,
  CHANGE_LAYER = 4,
  CHANGE_STAGE_SPEED = 5,
  CHANGE_STAGE_COLLISION = 6,
  CHANGE_MAX_TARGET = 7,
  CHANGE_AREA_RANGE = 8,
  CHANGE_AREA_ANGLE = 9,
  CHANGE_COST = 10,
  RECOVER_COST = 11,
  RECOVER_USED_COST = 12,
  REDUCE_DEFAULT_COOLDOWN = 13,
  REDUCE_ACTIVE_COOLDOWN = 14,
  ENABLE_STAGE_BUFF = 15,
  ADD_STAGE_BUFF = 16,
  CHANGE_BUFF_AREA_RANGE = 17,
  CHANGE_BUFF_DURATION = 18,
  CHANGE_BUFF_STAT = 19,
  CHANGE_BUFF_STACK = 20,
  CHANGE_BUFF_PARAM = 21,
  CHANGE_BUFF_EXPIRED_ACTION = 22,
  CHANGE_CHAIN_RATIO = 23,
  CHANGE_ABNORMAL = 24,
  CHANGE_ABNORMAL_RATIO = 25,
  CHANGE_DAM_ATTR = 26,
  CHANGE_DAM_VALUE = 27,
  CHANGE_DAM_COEFFICIENT = 28,
  CHANGE_DAM_CRITICAL = 29,
  CHANGE_DAM_CRITICAL_RATE = 30,
  CHANGE_ATTACK_STAGE_SPEED = 31,
  CHANGE_STACK_CHARGE_TIME = 32,
  CHANGE_STACK_MAX_COUNT = 33,
  CHANGE_TARGETING = 34,
  CHANGE_MIN_RANGE = 35,
  CHANGE_MAX_RANGE = 36,
  CHANGE_PUSH_INFO = 37,
  CHANGE_PARTS_ATTACK_ATTR = 38,
  CHANGE_SKILL_CHAIN_INFO = 39,
  CHANGE_SKILL_CHAIN_DELAY = 40,
  CHANGE_BEHIT_MOVE_INFO = 41,
  ADD_BUFF_STAT = 42,
  ADD_CHAIN_SKILL_EFFECT = 43,
  REMOVE_CHAIN_SKILL_EFFECT = 44,
  ADD_CHAIN_COMBAT_EFFECT = 45,
  REMOVE_CHAIN_COMBAT_EFFECT = 46,
  CHANGE_SKILL_EFFECT_BONUS = 47,
  CHANGE_SKILL_EFFECT_AI_POINT = 48,
  CHANGE_DAM_ADDEND = 49,
  CHANGE_HITTED = 50,
  CHANGE_SKILL_MOVE_SPEED = 51,
  ADD_SKILL_BUFF = 52,
  CHANGE_SKILL_BONUS = 53,
  CHANGE_SKILL_NORMAL_INFO = 54,
  CHANGE_SKILL_INVISIBILITY = 55,
  CHANGE_SKILL_CONSTRAINT = 56,
  CHANGE_SKILL_BOOK_TYPE = 57,
  CHANGE_PROJECTION_SKILL_EFFECT_ID = 58,
  CHANGE_PUSH_PVP_INFO = 59,
  CHANGE_FORCED_CRITICAL = 60,
  CHANGE_INSTANCE_SKILL_EFFECT_INFO = 61,
  CHANGE_SKILL_START_STAGE = 62,
  CHANGE_SKILL_EFFECT_DIR_TARGET = 63,
  CHANGE_STAGE_DIR_RATE = 64,
  CHANGE_PROJECTION = 65,
  CHANGE_SKILL_VIEW = 66,
  CHANGE_PROJECTILE_SPEED = 67,
  CHANGE_PROJECTILE_DIST = 68,
  CHANGE_PROJECTILE_RESOURCESCALE = 69,
  CHANGE_PROJECTILE_MAX_TARGET_HIT_COUNT = 70,
  CHANGE_SUMMON_TRAP_LIFETIME = 71,
  CHANGE_SUMMON_TRAP_DESTROY_DELAYTIME = 72,
  CHANGE_SUMMON_TRAP_REACT_INFO = 73,
  CHANGE_SUMMON_TRAP_INVOKE_EFFECT = 74,
  CHANGE_SUMMON_TRAP_COUNT = 75,
  ENABLE_IDENTITY_EVENT = 76,
  CHANGE_IDENTITY_PROC_VALUE = 77,
  CHANGE_SKILL_EFFECT_IDENTITY_PROC_INFO = 78,
  CHANGE_IDENTITY_PROC_PVP_VALUE = 79,
  CHANGE_SKILL_EFFECT_IDENTITY_PROC_PVP_INFO = 80,
  CHANGE_SKILL_EFFECT_IDENTITY_PROC_REPLACE_INFO = 81,
  CHANGE_SKILL_EFFECT_IDENTITY_PROC_REPLACE_PVP_INFO = 82,
  SWAP_CHAIN_SKILL_EFFECT = 83,
  SWAP_CHAIN_COMBAT_EFFECT = 84,
  CHANGE_CHARGE_SCALE = 85,
  CHANGE_SUMMON_NPC_ID = 86,
  CHANGE_SUMMON_NPC_SIGHT_RANGE = 87,
  CHANGE_SUMMON_NPC_PURSUIT_RANGE = 88,
  CHANGE_SUMMON_NPC_WALK_MOVESPEED = 89,
  CHANGE_SUMMON_NPC_BATTLE_MOVESPEED = 90,
  CHANGE_SUMMON_NPC_LIFE_TIME = 91,
  CHANGE_SUMMON_NPC_AI_INDEX = 92,
  CHANGE_SUMMON_NPC_INVINCIBLE_DURATION = 93,
  CHANGE_SUMMON_NPC_ACQUIRE_IDENTITY = 94,
  CHANGE_SUMMON_NPC_SKILL_ID = 95,
  CHANGE_SUMMON_NPC_DIE_SKILL_ID = 96,
  CHANGE_SUMMON_NPC_DESTROY_SKILL_ID = 97,
  CHANGE_SUMMON_NPC_SPAWN_BUFF_ID = 98,
  CHANGE_SUMMON_NPC_COUNT = 99,
  CHANGE_SUMMON_NPC_STAT = 100,
  CHANGE_SUMMON_NPC_THREAT_POINT = 101,
  CHANGE_SUMMON_NPC_SKILL_USABLE_TICK = 102,
  CHANGE_SUMMON_NPC_SKILL_USE_ORDER = 103,
  CHANGE_COMBAT_EFFECT_ARG = 104,
  CHANGE_SKILL_EFFECT_COST = 105,
  CHANGE_ACCUMULATE_DAM_RATE = 106,
  CHANGE_PROJECTILE_BANK_DATA_ADDEND = 107,
  CHANGE_IDENTITY_CATEGORY = 108,
  CHANGE_SKILL_SLOT_VISIBLE_EFFECT = 109,
  CHANGE_ATTACK_MASK = 110,
  CHANGE_AIM_TARGET_MAX_RANGE = 111,
  CHANGE_MULTI_HIT = 112,
  CHANGE_AWAKENING_DAM_VALUE = 113
}
// ...
export const enum EDamageAttr {
  NONE = 0,
  FIRE = 1,
  ICE = 2,
  ELECTRICITY = 3,
  WIND = 4,
  EARTH = 5,
  DARK = 6,
  HOLY = 7
}

export const enum ESkillEffectChainType {
  NONE = 0,
  HIT = 1,
  CRITICAL_HIT = 2
}

export const enum ESkillEffectChainActor {
  NONE = 0,
  SELF = 1,
  TARGET = 2
}